+++ src/utils/constants.js
@@ -1,0 +1,3 @@
+exports.GENESIS =
+  '0x4920616d207665727920616e6772792c20627574206974207761732066756e21';
+exports.EMPTY_ADDRESS = '0x0000000000000000000000000000000000000000';
