+++ src/utils/getSlotsByAddr.js
@@ -1,0 +1,5 @@
+const addrCmp = require('./addrCmp');
+
+module.exports = (slots, address) => {
+  return slots.filter(slot => addrCmp(slot.signer, address));
+};
